requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
python-dotenv>=1.0.0
groq>=0.3.0 
Pillow>=11.1.0
